resolve crossTriggerController's conflict

+++ business-component/src/main/java/com/kii/beehive/business/service/ThingMonitorService.java
@@ -1,16 +1,18 @@
 package com.kii.beehive.business.service;
 
 
-import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.kii.beehive.business.ruleengine.TriggerManager;
+import com.kii.beehive.portal.auth.AuthInfoStore;
 import com.kii.beehive.portal.common.utils.ThingIDTools;
 import com.kii.beehive.portal.jdbc.dao.GlobalThingSpringDao;
 import com.kii.beehive.portal.jdbc.dao.UserNoticeDao;
@@ -27,6 +29,7 @@ import com.kii.extension.ruleengine.store.trigger.TagSelector;
 import com.kii.extension.ruleengine.store.trigger.TriggerRecord;
 import com.kii.extension.ruleengine.store.trigger.groups.SummaryFunctionType;
 import com.kii.extension.ruleengine.store.trigger.target.CallBusinessFunction;
+import com.kii.extension.sdk.service.AbstractDataAccess;
 
 @Component("thingMonitorService")
 public class ThingMonitorService {
@@ -41,7 +44,7 @@ public class ThingMonitorService {
 	private ThingStatusMonitorDao  monitorDao;
 	
 	@Autowired
-	private GlobalThingSpringDao thingTagService;
+	private GlobalThingSpringDao thingDao;
 
 	@Autowired
 	private NoticeMsgQueue queue;
@@ -62,7 +65,7 @@ public class ThingMonitorService {
 		
 		ThingIDTools.ThingIDCombine ids=ThingIDTools.splitFullKiiThingID(thingID);
 
-		GlobalThingInfo th=thingTagService.getThingByFullKiiThingID(ids.kiiAppID,ids.kiiThingID);
+		GlobalThingInfo th=thingDao.getThingByFullKiiThingID(ids.kiiAppID,ids.kiiThingID);
 		notice.setTitle(th.getVendorThingID());
 		
 		monitor.getNoticeList().forEach(id->{
@@ -74,58 +77,72 @@ public class ThingMonitorService {
 		
 	}
 
-	public void addMonitor(ThingStatusMonitor  monitor){
-
-
-		TriggerRecord trigger=getTrigger(monitor);
+	public String addMonitor(ThingStatusMonitor  monitor){
+		
+		
+		monitor.getNoticeList().add(AuthInfoStore.getUserID());
+		
+		String monitorID=monitorDao.addEntity(monitor).getObjectID();
+		
+		List<Long> ids=thingDao.getThingsByVendorThings(monitor.getVendorThingIDList()).stream().map(GlobalThingInfo::getId).collect(Collectors.toList());
+		
+		monitor.setId(monitorID);
+		TriggerRecord trigger=getTrigger(monitor,ids);
 		
 		String triggerID=creator.createTrigger(trigger).getTriggerID();
 		
 		monitor.setRelationTriggerID(triggerID);
 		
-		monitorDao.addEntity(monitor);
-
+		monitorDao.updateEntity(Collections.singletonMap("relationTriggerID",triggerID),monitorID);
+		
+		return monitorID;
 	}
 	
 	public void removeMonitor(String id){
 		
 		ThingStatusMonitor monitor=monitorDao.getObjectByID(id);
 		
-		creator.deleteTrigger(monitor.getRelationTriggerID());
-		
-		monitorDao.updateEntity(Collections.singletonMap("status", ThingStatusMonitor.MonitorStatus.deleted),id);
+		if(monitor.getStatus()!= ThingStatusMonitor.MonitorStatus.deleted) {
+			
+			creator.deleteTrigger(monitor.getRelationTriggerID());
+			
+			monitorDao.updateEntity(Collections.singletonMap("status", ThingStatusMonitor.MonitorStatus.deleted), id);
+		}
 	}
 	
 	public void enableMonitor(String id){
 		
 		ThingStatusMonitor monitor=monitorDao.getObjectByID(id);
-		
-		creator.enableTrigger(monitor.getRelationTriggerID());
-		
-		monitorDao.updateEntity(Collections.singletonMap("status", ThingStatusMonitor.MonitorStatus.enable),id);
-		
-		
+		if(monitor.getStatus()!= ThingStatusMonitor.MonitorStatus.disable) {
+			creator.enableTrigger(monitor.getRelationTriggerID());
+			monitorDao.updateEntity(Collections.singletonMap("status", ThingStatusMonitor.MonitorStatus.enable), id);
+		}
 	}
 	
 	public void disableMonitor(String id){
 		ThingStatusMonitor monitor=monitorDao.getObjectByID(id);
 		
-		creator.disableTrigger(monitor.getRelationTriggerID());
-		
-		monitorDao.updateEntity(Collections.singletonMap("status", ThingStatusMonitor.MonitorStatus.disable),id);
-		
+		if(monitor.getStatus()!= ThingStatusMonitor.MonitorStatus.enable) {
+			
+			creator.disableTrigger(monitor.getRelationTriggerID());
+			
+			monitorDao.updateEntity(Collections.singletonMap("status", ThingStatusMonitor.MonitorStatus.disable), id);
+		}
 	}
 	
+	@Transactional
 	public void updateMonitor(ThingStatusMonitor monitor){
 		
 		
 		ThingStatusMonitor oldMonitor=monitorDao.getObjectByID(monitor.getId());
 		
-		if(!oldMonitor.getThings().equals(monitor.getThings())||
+		if(!oldMonitor.getVendorThingIDs().equals(monitor.getVendorThingIDs())||
 			!oldMonitor.getExpress().equals(monitor.getExpress())||
 			!oldMonitor.getCondition().equals(monitor.getCondition())) {
 			
-			TriggerRecord newTrigger=getTrigger(monitor);
+			List<Long> ids=thingDao.getThingsByVendorThings(monitor.getVendorThingIDList()).stream().map(GlobalThingInfo::getId).collect(Collectors.toList());
+			
+			TriggerRecord newTrigger=getTrigger(monitor,ids);
 			newTrigger.setTriggerID(oldMonitor.getRelationTriggerID());
 			creator.updateTrigger(newTrigger);
 			
@@ -135,17 +152,23 @@ public class ThingMonitorService {
 	}
 	
 	
-	private MultipleSrcTriggerRecord  getTrigger(ThingStatusMonitor monitor){
+	
+	private MultipleSrcTriggerRecord  getTrigger(ThingStatusMonitor monitor,List<Long> ids){
 		
 		MultipleSrcTriggerRecord record=new MultipleSrcTriggerRecord();
 		
-		record.setUsedByWho(TriggerRecord.UsedByType.Sys_monitor);
+		record.setUsedByWho(TriggerRecord.UsedByType.User_monitor);
 		
-		record.setName("mon"+monitor.getName());
+		record.setUserID(-999l);
+
+		record.fillCreator(monitor.getId());
+		
+		record.setName("mon"+monitor.getCreator()+"_"+monitor.getName());
 		GroupSummarySource source=new GroupSummarySource();
 		source.setFunction(SummaryFunctionType.objCol);
 		TagSelector src=new TagSelector();
-		src.setThingList(new ArrayList<>(monitor.getThings()));
+		
+		src.setThingList(ids);
 		source.setSource(src);
 		
 		Express express=new Express();
@@ -157,7 +180,7 @@ public class ThingMonitorService {
 		record.addSource("one", source);
 		
 		RuleEnginePredicate predicate=new RuleEnginePredicate();
-		predicate.setExpress("$p{one}!=$h{one}");
+		predicate.setExpress("$p:c{one}!=$h:c{one}");
 		record.setPredicate(predicate);
 		
 		
@@ -166,12 +189,15 @@ public class ThingMonitorService {
 		target.setFunctionName("addNotifiction");
 		target.setParamArrays("monitorID","currThing","thingStatus","sign");
 		
+		
 		record.addTarget(target);
 		
 		record.addTargetParam("currThing","$e{sys.curr.currThing}");
-		record.addTargetParam("monitorID",monitor.getId());
+		record.addTargetParam("monitorID","'"+monitor.getId()+"'");
 		record.addTargetParam("thingStatus","$e{runtime.currStatus}");
 		record.addTargetParam("sign","$p:c{one}.contains($e{sys.curr.currThing})");
+		
+		record.setRecordStatus(TriggerRecord.StatusType.enable);
 		
 		return record;
 	}
@@ -180,6 +206,14 @@ public class ThingMonitorService {
 		
 		return monitorDao.getObjectByID(monitorID);
 	}
+	
+	
+	public List<ThingStatusMonitor> queryMonitor(ThingStatusMonitorDao.MonitorQuery query,AbstractDataAccess.KiiBucketPager pager){
+		
+		return monitorDao.getMonitorsByQuery(query,pager);
+		
+	}
+	
 
 /*
 {
